requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
chromadb==1.0.12
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
db-dtypes==1.4.3
distro==1.9.0
docstring_parser==0.16
durationpy==0.10
fastapi==0.115.9
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.97.0
google-cloud-bigquery==3.34.0
google-cloud-bigquery-storage==2.32.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.21.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
jieba==0.42.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.26
langchain-core==0.3.66
langchain-google-community==2.0.7
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.8
langsmith==0.3.45
markdown-it-py==3.0.0
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.5.0
mypy_extensions==1.1.0
numpy==2.3.0
oauthlib==3.3.0
onnxruntime==1.19.2
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pandas-gbq==0.29.1
pillow==11.2.1
posthog==5.0.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.27.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydata-google-auth==1.9.1
Pygments==2.19.1
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.45.3
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.21.1
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
